removed old reviews component + resolved test
