--- src/components/work/_reviews/Reviews.module.css
@@ -1,54 +1,0 @@
-.reviews {
-  position: relative;
-  height: auto;
-}
-
-.only_content {
-  display: contents;
-}
-
-.slider > div > div {
-  scroll-snap-align: start;
-}
-
-.slider .trio {
-  scroll-snap-align: start;
-  display: flex;
-  flex-direction: row;
-}
-
-/* Slider */
-
-.SlideWrapper {
-  height: auto;
-  padding-right: var(--pt3);
-  align-self: flex-start;
-}
-
-/* Skeleton */
-
-.skeleton .button {
-  display: none;
-}
-
-.skeleton.custom div div div i div,
-.skeleton.custom div > div div,
-.skeleton.custom div span div div div {
-  background-color: var(--parchment);
-}
-
-.skeleton.custom div div::before {
-  background: linear-gradient(
-    to right,
-    transparent 0%,
-    var(--chamois) 50%,
-    transparent 100%
-  );
-}
-
-/* Medium devices (tablets, 768px and up) */
-@media (min-width: 992px) {
-  .SlideWrapper {
-    max-width: unset;
-  }
-}
