--- src/components/work/_reviews/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Reviews.js";
