--- src/components/work/_reviews/types/external/ExternalReview.js
@@ -1,213 +1,0 @@
-import PropTypes from "prop-types";
-import Col from "react-bootstrap/Col";
-
-import { cyKey } from "@/utils/trim";
-
-import Text from "@/components/base/text";
-import Link from "@/components/base/link";
-import Title from "@/components/base/title";
-import Icon from "@/components/base/icon";
-import Rating from "@/components/base/rating";
-import Translate from "@/components/base/translate";
-
-import { dateToShortDate, numericToISO } from "@/utils/datetimeConverter";
-
-import styles from "./ExternalReview.module.css";
-
-/**
- * The Component function
- *
- * @param {obj} props
- * See propTypes for specific props and types
- *
- * @returns {component}
- */
-export function ExternalReview({
-  className = "",
-  data = [],
-  onFocus = null,
-  skeleton = false,
-}) {
-  // Translate Context
-  const context = { context: "reviews" };
-
-  const volume =
-    data.hostPublication?.issue ||
-    (data.recordCreationDate &&
-      dateToShortDate(numericToISO(data.recordCreationDate), "d. "));
-
-  return (
-    <Col
-      xs={11}
-      sm={6}
-      md={4}
-      className={`${styles.litteratursiden} ${className}`}
-      data-cy={cyKey({ prefix: "review", name: "external" })}
-    >
-      <div>
-        {data.hostPublication?.title && (
-          <div className={styles.media}>
-            <Title type="title4" skeleton={skeleton}>
-              {data.hostPublication?.title}
-            </Title>
-          </div>
-        )}
-        <div className={styles.row}>
-          {data.creators?.length > 0 && (
-            <div className={styles.left}>
-              <Text type="text3" skeleton={skeleton} lines={1}>
-                {Translate({ context: "general", label: "by" })}
-              </Text>
-            </div>
-          )}
-
-          <div className={styles.right}>
-            {data.creators?.length > 0 && (
-              <div className={styles.author}>
-                {!skeleton && (
-                  <Text type="text2">
-                    {data.creators?.map((c) => c.display).join(", ")}
-                  </Text>
-                )}
-                <div className={styles.date}>
-                  {!skeleton && volume && <Text type="text3">{volume}</Text>}
-                </div>
-              </div>
-            )}
-
-            {data?.creators?.length === 0 && !skeleton && volume && (
-              <>
-                <Text type="text3" className={styles.date}>
-                  Udgivet: {volume}
-                </Text>
-                {data?.physicalDescriptions?.length > 0 && (
-                  <Text type="text3" className={styles.date}>
-                    {data?.physicalDescriptions
-                      ?.map((desc) => desc.summary)
-                      .join(", ")}
-                  </Text>
-                )}
-              </>
-            )}
-
-            {data?.review?.rating && (
-              <div className={styles.rating}>
-                <Rating rating={data?.review?.rating} skeleton={skeleton} />
-              </div>
-            )}
-          </div>
-        </div>
-
-        {data.access?.map((access) => {
-          if (access.__typename === "AccessUrl") {
-            if (access.url && access.url !== "") {
-              const shouldUseAlternateText =
-                access.url?.includes("https://moreinfo");
-              return (
-                <div className={styles.url} key={access.url}>
-                  <Icon
-                    src="chevron.svg"
-                    size={{ w: 2, h: "auto" }}
-                    skeleton={skeleton}
-                    alt=""
-                  />
-                  <Link
-                    href={access.url}
-                    target="_blank"
-                    onFocus={onFocus}
-                    disabled={!access.url}
-                    border={{ top: false, bottom: { keepVisible: true } }}
-                  >
-                    <Text type="text2" skeleton={skeleton}>
-                      {Translate({
-                        ...context,
-                        label: shouldUseAlternateText
-                          ? "alternateReviewLinkText"
-                          : "reviewLinkText",
-                      })}
-                    </Text>
-                  </Link>
-                </div>
-              );
-            }
-          }
-        })}
-      </div>
-    </Col>
-  );
-}
-
-/**
- * Function to return skeleton (Loading) version of the Component
- *
- * @param {obj} props
- *  See propTypes for specific props and types
- *
- * @returns {component}
- */
-export function ExternalReviewSkeleton(props) {
-  const data = {
-    pid: "Some pid",
-    creators: [
-      {
-        display: "Some creator",
-      },
-    ],
-    access: [
-      {
-        __typename: "AccessUrl",
-        origin: "Some domain",
-        url: "http://www.some-url.dk",
-        note: "Some note",
-        loginRequired: false,
-        type: "RESOURCE",
-      },
-      {
-        __typename: "DigitalArticleService",
-        issn: "Some issn",
-      },
-    ],
-    hostPublication: {
-      title: "Some title",
-      issue: "Nr. 1 (2006)",
-    },
-    recordCreationDate: "20061120",
-    review: {
-      rating: "5/6",
-    },
-  };
-
-  return (
-    <ExternalReview
-      {...props}
-      data={data}
-      className={`${props.className} ${styles.skeleton}`}
-      skeleton={true}
-    />
-  );
-}
-
-/**
- *  Default export function of the Component
- *
- * @param {obj} props
- * See propTypes for specific props and types
- *
- * @returns {component}
- */
-export default function Wrap(props) {
-  const { skeleton } = props;
-
-  if (skeleton) {
-    return <ExternalReviewSkeleton />;
-  }
-
-  return <ExternalReview {...props} />;
-}
-
-// PropTypes for component
-Wrap.propTypes = {
-  workId: PropTypes.string,
-  type: PropTypes.string,
-  skeleton: PropTypes.bool,
-};
