--- src/components/work/_reviews/types/external/ExternalReview.module.css
@@ -1,64 +1,0 @@
-.litteratursiden {
-  padding-bottom: var(--pt1);
-  padding-top: var(--pt7);
-  display: flex;
-  align-self: center;
-  justify-content: center;
-}
-
-.litteratursiden > div {
-  width: 100%;
-}
-
-.litteratursiden a {
-  text-decoration: none;
-}
-
-.litteratursiden a:focus {
-  color: var(--blue);
-}
-
-.media {
-  padding: 0 0 var(--pt2) 0;
-}
-
-.row {
-  display: flex;
-}
-
-.left {
-  min-width: 24px;
-  padding: 2px 0 0 0;
-  flex-shrink: 2;
-}
-
-.right {
-  padding: 0;
-}
-
-.author {
-  padding: 0;
-}
-
-.rating {
-  padding: var(--pt2) 0 0 0;
-}
-
-.url {
-  padding: var(--pt3) 0 0 0;
-  display: inline-flex;
-  align-items: center;
-}
-
-.url a {
-  margin: 0 0 0 var(--pt2);
-}
-.alternateUrl {
-  padding-top: var(--pt2);
-}
-
-/* Skeleton */
-
-.skeleton .author p {
-  display: block;
-}
