--- src/components/work/_reviews/types/external/index.js
@@ -1,1 +1,0 @@
-export { default } from "./ExternalReview.js";
