--- src/components/work/_reviews/types/infomedia/index.js
@@ -1,1 +1,0 @@
-export { default } from "./InfomediaReview.js";
