--- src/components/work/_reviews/types/material/index.js
@@ -1,1 +1,0 @@
-export { default } from "./MaterialReview.js";
